+++ lib/prelude/src/collections.rs
@@ -1,0 +1,6 @@
+//! This module exports collections which are popular enough to be available everywhere.
+
+pub use smallvec::SmallVec;
+pub use std::collections::BTreeMap;
+pub use std::collections::HashMap;
+pub use std::collections::HashSet;
