+++ lib/prelude/src/lib.rs
@@ -9,26 +9,27 @@
 #![feature(specialization)]
 #![feature(trait_alias)]
 
+pub mod collections;
 pub mod macros;
 pub mod option;
 pub mod phantom;
-pub mod std_reexports;
 pub mod reference;
+pub mod std_reexports;
 pub mod string;
 pub mod tp;
 pub mod wrapper;
 
+pub use collections::*;
 pub use macros::*;
 pub use option::*;
 pub use phantom::*;
-pub use std_reexports::*;
 pub use reference::*;
+pub use std_reexports::*;
 pub use string::*;
 pub use tp::*;
 pub use wrapper::*;
 
 pub use boolinator::Boolinator;
-
 pub use derivative::Derivative;
 pub use derive_more::*;
 pub use enclose::enclose;
@@ -39,7 +40,6 @@ pub use lazy_static::lazy_static;
 pub use num::Num;
 pub use paste;
 pub use shrinkwraprs::Shrinkwrap;
-pub use smallvec::SmallVec;
 
 
 
@@ -65,78 +65,6 @@ impl<T:?Sized> CloneRef for Rc<T> {
     }
 }
 
-
-
-// ===================
-// === WithPhantom ===
-// ===================
-
-/// A wrapper adding a phantom type to a structure.
-#[derive(Derivative)]
-#[derive(Shrinkwrap)]
-#[shrinkwrap(mutable)]
-#[derivative(Clone   (bound="T:Clone"))]
-#[derivative(Default (bound="T:Default"))]
-#[derivative(Debug   (bound="T:Debug"))]
-pub struct WithPhantom<T, P=()> {
-    #[shrinkwrap(main_field)]
-    pub without_phantom: T,
-    phantom: PhantomData<P>
-}
-
-impl<T, P> WithPhantom<T, P> {
-    pub fn new(without_phantom: T) -> Self {
-        let phantom = PhantomData;
-        Self { without_phantom, phantom }
-    }
-}
-
-
-
-// ==========================
-// === PhantomConversions ===
-// ==========================
-
-/// A utility for easy driving of type-level computations from value level. Often we've got some
-/// type level relations, like a few singleton types, and for each such type we've got an associated
-/// value. For example, we can define types `Int` and `Float` and associate with them
-/// `WebGlContext::Int` and `WebGlContext::Float` constants encoded as `GlEnum`. In order to convert
-/// `Int` or `Float` to the `GlEnum` we do not need the instance of the types, only the information
-/// what type it was. So we can define:
-///
-/// ```compile_fail
-/// impl From<PhantomData<Int>> for u32 {
-///     from(_:PhantomData<Int>>) {
-///         GlEnum(WebGlContext::Int)
-///     }
-/// }
-/// ```
-///
-/// And use it like:
-///
-/// ```compile_fail
-/// let val = GlEnum::from(PhantomData::<Int>)
-/// ```
-///
-/// Using this utility we can always write the following code instead:
-///
-/// ```compile_fail
-/// let val = GlEnum::phantom_from::<Int>()
-/// ```
-pub trait PhantomConversions: Sized {
-    fn phantom_into<P>() -> P where Self:PhantomInto<P> {
-        PhantomData::<Self>.into()
-    }
-    fn phantom_from<P:PhantomInto<Self>>() -> Self {
-        PhantomData::<P>.into()
-    }
-}
-impl<T> PhantomConversions for T {}
-
-/// Like `Into` but for phantom types.
-pub trait PhantomInto<T> = where PhantomData<Self>: Into<T>;
-
-
 /// Provides method `to`, which is just like `into` but allows fo superfish syntax.
 pub trait ToImpl: Sized {
     fn to<P>(self) -> P where Self:Into<P> {
@@ -144,7 +72,6 @@ pub trait ToImpl: Sized {
     }
 }
 impl<T> ToImpl for T {}
-
 
 
 // TODO
@@ -161,8 +88,6 @@ where T:nalgebra::Scalar, R:nalgebra::DimName, C:nalgebra::DimName {
         }
     }
 }
-
-
 
 #[macro_export]
 macro_rules! clone_boxed {
@@ -186,94 +111,5 @@ macro_rules! clone_boxed {
     }}
 }
 
-
-// ===================
-// === WithContent2 ===
-// ===================
-
-pub trait WithContent2 {
-    type Content;
-    fn with_content<F:FnOnce(&Self::Content)->T,T>(&self, f:F) -> T;
-}
-
-impl<T:Deref> WithContent2 for T
-    where <T as Deref>::Target: WithContent2 {
-    type Content = <<T as Deref>::Target as WithContent2>::Content;
-    default fn with_content<F:FnOnce(&Self::Content)->R,R>(&self, f:F) -> R {
-        self.deref().with_content(f)
-    }
-}
-
-
-
-// =============
-// === Value ===
-// =============
-
-/// Defines relation between types and values, like between `True` and `true`.
-pub trait KnownTypeValue {
-
-    /// The value-level counterpart of this type-value.
-    type Value;
-
-    /// The value of this type-value.
-    fn value() -> Self::Value;
-}
-
-pub type TypeValue<T> = <T as KnownTypeValue>::Value;
-
-
-
-// =======================
-// === Type-level Bool ===
-// =======================
-
-/// Type level `true` value.
-#[derive(Clone,Copy,Debug)]
-pub struct True {}
-
-/// Type level `false` value.
-#[derive(Clone,Copy,Debug)]
-pub struct False {}
-
-impl KnownTypeValue for True {
-    type Value = bool;
-    fn value() -> Self::Value {
-        true
-    }
-}
-
-impl KnownTypeValue for False {
-    type Value = bool;
-    fn value() -> Self::Value {
-        false
-    }
-}
-
 /// Alias for `for<'t> &'t Self : Into<T>`.
 pub trait RefInto<T> = where for<'t> &'t Self : Into<T>;
-
-
-
-// =============
-// === Owned ===
-// =============
-
-pub trait AsOwned {
-    type Owned;
-}
-
-impl<T> AsOwned for &T {
-    type Owned = T;
-}
-
-pub type Owned<T> = <T as AsOwned>::Owned;
-
-pub trait IntoOwned = AsOwned + Into<Owned<Self>>;
-
-
-
-/// Placeholder type used to represent any value. It is useful to define type-level relations like
-/// defining an unit with any quantity, let it be distance or mass.
-#[derive(Clone,Copy,Debug,PartialEq)]
-pub struct Anything {}
